+++ api-module/src/main/kotlin/com/cj/smartworker/dataaccess/fcm/adapter/FcmHistoryPersistenceAdapter.kt
@@ -17,8 +17,8 @@ internal class FcmHistoryPersistenceAdapter(
         reporter: Member,
         admins: Set<Member>,
         createdAt: LocalDateTime,
-        x: Int,
-        y: Int,
+        x: Float,
+        y: Float,
         emergency: Emergency,
     ) {
         val fcmHistoryJpaEntity = FcmHistoryJpaEntity(
